Updated server.js
* Created cgi-bin and equations_flask_app.cgi

This has the script that will be used for getting data from fields (Username, password, e-mail). It does not have other data storing methods at the moment.

* Created server.js

Has turn increment logic. Not much else can be done without other data.

* Turns, adding players, and moving

This code adds a logic for how players will join and play the game.

* Updated server.js

Made a few changes and added turn changes when a player leaves or makes their move.

* Added database, check validity, and finding game by id

This makes use of a database. It does not make use of the one that we currently have. Connectivity of that will have to come later.