Smingolellihtml5 (#38)
* Fixed Server Code

There were a couple typos and placement issues that needed to be resolved.

* Created cgi-bin and equations_flask_app.cgi

This has the script that will be used for getting data from fields (Username, password, e-mail). It does not have other data storing methods at the moment.

* Added a line that stopped server crash.

This line will more or less deal with index OOB issues.

* Added actions and report_choice

The server will report choices made by a user, but this method has been commented out.

* Simplify Server using Pusher

In order to make things (hopefully) easier I decided to use a mix of Flask and Pusher to make the multiplayer process easier. I'm still looking into how it works.

* Created server.js

Has turn increment logic. Not much else can be done without other data.

* Turns, adding players, and moving

This code adds a logic for how players will join and play the game.

* Updated server.js

Made a few changes and added turn changes when a player leaves or makes their move.

* Added database, check validity, and finding game by id

This makes use of a database. It does not make use of the one that we currently have. Connectivity of that will have to come later.

* Added valid data system

This will now validate data before a player can create or join a game. One problem I see, but I don't have an answer to, is that I don't know how to check if a room is full or how many players are in it.

* Fixed moving and withdrawing

Fixed lines that called the game variable, as it was not initialised.

* Updated server.js, added package.json, and node.js

* Made Equations_Controller.java and tests

This is the controller. It interacts calls with the model, but when relevant, the view.